refactor: move filter into dedicated file
