+++ app.js
@@ -1,5 +1,6 @@
 const { data } = require('./data')
-const { count } = require('./count')
+const { filterCountries } = require('./commands/filter')
+const { count } = require('./commands/count')
 const { serializeArguments, getArgumentValue } = require('./utils')
 
 const main = () => {
@@ -16,40 +17,8 @@ const main = () => {
     }
 }
 
-const filterCountries = (countryList, filter) => {
-    const filterCountryList = countryList.reduce((list, country) => {
-        const filterPeopleList = filterPeople(country.people, filter)
-        if (filterPeopleList.length) {
-            list.push({ ...country, people: filterPeopleList })
-        }
-        return list
-    }, [])
-    return filterCountryList
-}
-
-const filterPeople = (peopleList, filter) => {
-    const filteredPeopleList = peopleList.reduce((list, people) => {
-        const filteredAnimalList = filterAnimals(people.animals, filter)
-        if (filteredAnimalList.length) {
-            list.push({ ...people, animals: filteredAnimalList })
-        }
-        return list
-    }, [])
-    return filteredPeopleList
-}
-
-const filterAnimals = (animalList, filter) => {
-    const filteredAnimalList = animalList.filter((animal) =>
-        animal.name.includes(filter)
-    )
-    return filteredAnimalList
-}
-
 main()
 
 module.exports = {
     main,
-    filterCountries,
-    filterPeople,
-    filterAnimals,
 }
