+++ commands/count.js
