+++ commands/count.test.js
